refactor(backend): Decompose friends service and routes into focused modules
Extract large monolithic files into focused, testable modules following
a delegation pattern that maintains backward compatibility:

Services (apps/backend/src/services/friends/):
- search.service.ts: Full-text search with faceted filtering
- relationship.service.ts: Friend relationship management
- network-graph.service.ts: Network visualization data
- base/sub-resource.service.ts: Generic base class for sub-resources
- sub-resources/: Phone, email, address, url, date, social-profile,
  professional-history, and met-info services

Routes (apps/backend/src/routes/friends/):
- core.routes.ts: Friend CRUD operations
- search.routes.ts: All search endpoints
- dashboard.routes.ts: Upcoming dates, network graph
- organization.routes.ts: Photos, circles, favorites, archive
- sub-resources/: Individual route files for each sub-resource type

Key improvements:
- Reduced friends.service.ts from 2,302 to ~700 lines
- Reduced friends.ts routes from 2,084 to re-export only
- Added InvalidSearchParametersError for proper error handling
- Maintained 100% backward compatibility via re-exports
- All 200 integration tests pass
