+++ apps/backend/src/routes/friends/sub-resources/emails.routes.ts
@@ -1,0 +1,134 @@
+import { EmailInputSchema, type ErrorResponse } from '@freundebuch/shared/index.js';
+import * as Sentry from '@sentry/node';
+import { type } from 'arktype';
+import { Hono } from 'hono';
+import { getAuthUser } from '../../../middleware/auth.js';
+import { FriendsService } from '../../../services/friends/index.js';
+import type { AppContext } from '../../../types/context.js';
+import { toError } from '../../../utils/errors.js';
+import { isValidUuid } from '../../../utils/security.js';
+
+const app = new Hono<AppContext>();
+
+/**
+ * POST /api/friends/:id/emails
+ * Add an email address to a friend
+ */
+app.post('/', async (c) => {
+  const logger = c.get('logger');
+  const db = c.get('db');
+  const user = getAuthUser(c);
+  const friendId = c.req.param('id')!;
+
+  if (!isValidUuid(friendId)) {
+    return c.json<ErrorResponse>({ error: 'Invalid friend ID' }, 400);
+  }
+
+  try {
+    let body: unknown;
+    try {
+      body = await c.req.json();
+    } catch {
+      return c.json<ErrorResponse>({ error: 'Invalid JSON' }, 400);
+    }
+
+    const validated = EmailInputSchema(body);
+
+    if (validated instanceof type.errors) {
+      return c.json<ErrorResponse>({ error: 'Invalid request', details: validated }, 400);
+    }
+
+    const friendsService = new FriendsService(db, logger);
+    const email = await friendsService.addEmail(user.userId, friendId, validated);
+
+    if (!email) {
+      return c.json<ErrorResponse>({ error: 'Friend not found' }, 404);
+    }
+
+    return c.json(email, 201);
+  } catch (error) {
+    const err = toError(error);
+    logger.error({ err, friendId }, 'Failed to add email');
+    Sentry.captureException(err);
+    return c.json<ErrorResponse>({ error: 'Failed to add email' }, 500);
+  }
+});
+
+/**
+ * PUT /api/friends/:id/emails/:emailId
+ * Update an email address
+ */
+app.put('/:emailId', async (c) => {
+  const logger = c.get('logger');
+  const db = c.get('db');
+  const user = getAuthUser(c);
+  const friendId = c.req.param('id')!;
+  const emailId = c.req.param('emailId')!;
+
+  if (!isValidUuid(friendId) || !isValidUuid(emailId)) {
+    return c.json<ErrorResponse>({ error: 'Invalid ID' }, 400);
+  }
+
+  try {
+    let body: unknown;
+    try {
+      body = await c.req.json();
+    } catch {
+      return c.json<ErrorResponse>({ error: 'Invalid JSON' }, 400);
+    }
+
+    const validated = EmailInputSchema(body);
+
+    if (validated instanceof type.errors) {
+      return c.json<ErrorResponse>({ error: 'Invalid request', details: validated }, 400);
+    }
+
+    const friendsService = new FriendsService(db, logger);
+    const email = await friendsService.updateEmail(user.userId, friendId, emailId, validated);
+
+    if (!email) {
+      return c.json<ErrorResponse>({ error: 'Email not found' }, 404);
+    }
+
+    return c.json(email);
+  } catch (error) {
+    const err = toError(error);
+    logger.error({ err, friendId, emailId }, 'Failed to update email');
+    Sentry.captureException(err);
+    return c.json<ErrorResponse>({ error: 'Failed to update email' }, 500);
+  }
+});
+
+/**
+ * DELETE /api/friends/:id/emails/:emailId
+ * Delete an email address
+ */
+app.delete('/:emailId', async (c) => {
+  const logger = c.get('logger');
+  const db = c.get('db');
+  const user = getAuthUser(c);
+  const friendId = c.req.param('id')!;
+  const emailId = c.req.param('emailId')!;
+
+  if (!isValidUuid(friendId) || !isValidUuid(emailId)) {
+    return c.json<ErrorResponse>({ error: 'Invalid ID' }, 400);
+  }
+
+  try {
+    const friendsService = new FriendsService(db, logger);
+    const deleted = await friendsService.deleteEmail(user.userId, friendId, emailId);
+
+    if (!deleted) {
+      return c.json<ErrorResponse>({ error: 'Email not found' }, 404);
+    }
+
+    return c.json({ message: 'Email deleted successfully' });
+  } catch (error) {
+    const err = toError(error);
+    logger.error({ err, friendId, emailId }, 'Failed to delete email');
+    Sentry.captureException(err);
+    return c.json<ErrorResponse>({ error: 'Failed to delete email' }, 500);
+  }
+});
+
+export default app;
