+++ apps/backend/src/routes/friends/sub-resources/urls.routes.ts
@@ -1,0 +1,134 @@
+import { type ErrorResponse, UrlInputSchema } from '@freundebuch/shared/index.js';
+import * as Sentry from '@sentry/node';
+import { type } from 'arktype';
+import { Hono } from 'hono';
+import { getAuthUser } from '../../../middleware/auth.js';
+import { FriendsService } from '../../../services/friends/index.js';
+import type { AppContext } from '../../../types/context.js';
+import { toError } from '../../../utils/errors.js';
+import { isValidUuid } from '../../../utils/security.js';
+
+const app = new Hono<AppContext>();
+
+/**
+ * POST /api/friends/:id/urls
+ * Add a URL to a friend
+ */
+app.post('/', async (c) => {
+  const logger = c.get('logger');
+  const db = c.get('db');
+  const user = getAuthUser(c);
+  const friendId = c.req.param('id')!;
+
+  if (!isValidUuid(friendId)) {
+    return c.json<ErrorResponse>({ error: 'Invalid friend ID' }, 400);
+  }
+
+  try {
+    let body: unknown;
+    try {
+      body = await c.req.json();
+    } catch {
+      return c.json<ErrorResponse>({ error: 'Invalid JSON' }, 400);
+    }
+
+    const validated = UrlInputSchema(body);
+
+    if (validated instanceof type.errors) {
+      return c.json<ErrorResponse>({ error: 'Invalid request', details: validated }, 400);
+    }
+
+    const friendsService = new FriendsService(db, logger);
+    const url = await friendsService.addUrl(user.userId, friendId, validated);
+
+    if (!url) {
+      return c.json<ErrorResponse>({ error: 'Friend not found' }, 404);
+    }
+
+    return c.json(url, 201);
+  } catch (error) {
+    const err = toError(error);
+    logger.error({ err, friendId }, 'Failed to add URL');
+    Sentry.captureException(err);
+    return c.json<ErrorResponse>({ error: 'Failed to add URL' }, 500);
+  }
+});
+
+/**
+ * PUT /api/friends/:id/urls/:urlId
+ * Update a URL
+ */
+app.put('/:urlId', async (c) => {
+  const logger = c.get('logger');
+  const db = c.get('db');
+  const user = getAuthUser(c);
+  const friendId = c.req.param('id')!;
+  const urlId = c.req.param('urlId')!;
+
+  if (!isValidUuid(friendId) || !isValidUuid(urlId)) {
+    return c.json<ErrorResponse>({ error: 'Invalid ID' }, 400);
+  }
+
+  try {
+    let body: unknown;
+    try {
+      body = await c.req.json();
+    } catch {
+      return c.json<ErrorResponse>({ error: 'Invalid JSON' }, 400);
+    }
+
+    const validated = UrlInputSchema(body);
+
+    if (validated instanceof type.errors) {
+      return c.json<ErrorResponse>({ error: 'Invalid request', details: validated }, 400);
+    }
+
+    const friendsService = new FriendsService(db, logger);
+    const url = await friendsService.updateUrl(user.userId, friendId, urlId, validated);
+
+    if (!url) {
+      return c.json<ErrorResponse>({ error: 'URL not found' }, 404);
+    }
+
+    return c.json(url);
+  } catch (error) {
+    const err = toError(error);
+    logger.error({ err, friendId, urlId }, 'Failed to update URL');
+    Sentry.captureException(err);
+    return c.json<ErrorResponse>({ error: 'Failed to update URL' }, 500);
+  }
+});
+
+/**
+ * DELETE /api/friends/:id/urls/:urlId
+ * Delete a URL
+ */
+app.delete('/:urlId', async (c) => {
+  const logger = c.get('logger');
+  const db = c.get('db');
+  const user = getAuthUser(c);
+  const friendId = c.req.param('id')!;
+  const urlId = c.req.param('urlId')!;
+
+  if (!isValidUuid(friendId) || !isValidUuid(urlId)) {
+    return c.json<ErrorResponse>({ error: 'Invalid ID' }, 400);
+  }
+
+  try {
+    const friendsService = new FriendsService(db, logger);
+    const deleted = await friendsService.deleteUrl(user.userId, friendId, urlId);
+
+    if (!deleted) {
+      return c.json<ErrorResponse>({ error: 'URL not found' }, 404);
+    }
+
+    return c.json({ message: 'URL deleted successfully' });
+  } catch (error) {
+    const err = toError(error);
+    logger.error({ err, friendId, urlId }, 'Failed to delete URL');
+    Sentry.captureException(err);
+    return c.json<ErrorResponse>({ error: 'Failed to delete URL' }, 500);
+  }
+});
+
+export default app;
