+++ apps/backend/src/services/friends/index.ts
@@ -1,0 +1,23 @@
+// Re-export all friend-related services
+
+export type { SubResourceConfig, SubResourceServiceOptions } from './base/sub-resource.service.js';
+// Base service
+export { SubResourceService } from './base/sub-resource.service.js';
+export { FriendsService } from './friends.service.js';
+export type { NetworkGraphServiceOptions } from './network-graph.service.js';
+export { NetworkGraphService } from './network-graph.service.js';
+export type { RelationshipServiceOptions } from './relationship.service.js';
+export { RelationshipService } from './relationship.service.js';
+export type { SearchServiceOptions } from './search.service.js';
+export { createWildcardQuery, escapeLikePattern, SearchService } from './search.service.js';
+export { AddressService } from './sub-resources/address.service.js';
+export type { DateServiceOptions } from './sub-resources/date.service.js';
+export { DateService } from './sub-resources/date.service.js';
+export { EmailService } from './sub-resources/email.service.js';
+export type { MetInfoServiceOptions } from './sub-resources/met-info.service.js';
+export { MetInfoService } from './sub-resources/met-info.service.js';
+// Sub-resource services
+export { PhoneService } from './sub-resources/phone.service.js';
+export { ProfessionalHistoryService } from './sub-resources/professional-history.service.js';
+export { SocialProfileService } from './sub-resources/social-profile.service.js';
+export { UrlService } from './sub-resources/url.service.js';
